refactor yogogawa module
