+++ src/fractal_uzh_converters/__FRACTAL_MANIFEST__.json
@@ -2323,7 +2323,7 @@
       ],
       "docs_info": "### Purpose\n\n- Convert images acquired with a Yokogawa CellVoyager microscope to an OME-Zarr Plate.\n\n### Outputs\n\n- An OME-Zarr Plate.\n\n### Channels\n\n- Channel labels, colors and wavelength ids (`A{action}_C{channel}`) are read from the acquisition's `.mes` protocol file, whose name is recorded in the `.mrf`. Without a `.mes`, the label is the wavelength id.\n- To override them, `Advanced` \u2192 `Channels` is ordered by the **instrument channel number**: element 0 is `Ch1`, element 1 is `Ch2`, over the full range the protocol declares \u2014 not one entry per channel visible in the output. It must be at least as long as the highest channel the acquisition uses; the error names that number if it is too short.\n\n### Metadata\n\n- The acquisition's five plate-level vendor files are copied verbatim into `<plate>.zarr/metadata/`: the `.mlf` and `.mrf` under their fixed names, the `.mes`, `.wpi` and `.wpp` under the names recorded in the `.mrf`.\n- A file the acquisition does not ship is only a warning \u2014 CV8000 acquisitions routinely name a `.wpi` that was never written \u2014 and a name already taken by another acquisition becomes `<stem>_acq{id}<ext>` instead of overwriting.\n\n### Limitations\n\n- Unlike the CQ3K converter, this task does not support Z-image processing types.\n\n### Expected inputs\n\nThe following directory structure is expected:\n\n```text\nmy_acquisition/\n\u251c\u2500\u2500 MeasurementData.mlf      # Image measurement records (required)\n\u251c\u2500\u2500 MeasurementDetail.mrf    # Acquisition details and channel info (required)\n\u251c\u2500\u2500 protocol.mes             # Acquisition protocol (optional, sets channel names)\n\u251c\u2500\u2500 plate.wpi                # Plate definition (optional)\n\u251c\u2500\u2500 plate.wpp                # Plate product (optional)\n\u251c\u2500\u2500 image_001.png\n\u2514\u2500\u2500 ...\n```\n\nThe image file paths are referenced inside `MeasurementData.mlf` (with `.tif` extension) and can be in subdirectories relative to the acquisition directory. The actual files may use `.png` or `.tif` extension \u2014 select the matching extension via the `image_extension` parameter. The `.mes`, `.wpi` and `.wpp` filenames vary per acquisition and are read out of the `.mrf` \u2014 the names above are only examples.\n",
       "type": "converter_compound",
-      "executable_non_parallel": "cellvoyager/convert_cellvoyager_init_task.py",
+      "executable_non_parallel": "yokogawa/cellvoyager/convert_cellvoyager_init_task.py",
       "executable_parallel": "common/image_in_plate_compute_task.py",
       "meta_non_parallel": {
         "cpus_per_task": 1,
@@ -4632,7 +4632,7 @@
       ],
       "docs_info": "### Purpose\n\n- Convert images acquired with a Yokogawa CQ3K microscope to an OME-Zarr Plate.\n\n### Outputs\n\n- An OME-Zarr Plate.\n- One plate per Z-image processing type. Projections are suffixed `_MIP` (`Maximum`), `_MinIP` (`Minimum`) or `_SIP` (`Sum`), and coexist with the unsuffixed plate holding the raw Z slices. Channel labels are the same across all of them.\n- `Advanced` \u2192 `Z Processing` is one switch per output. Leave it unset to convert all of them; otherwise only `Z slices` is on by default, so enabling a projection adds to the raw stack rather than replacing it. An enabled kind the acquisition does not contain is a warning; enabling nothing, or nothing that the acquisition contains, is an error.\n\n### Channels\n\n- Channel labels, colors and wavelength ids (`A{action}_C{channel}`) are read from the acquisition's `.mes` protocol file, whose name is recorded in the `.mrf`. Without a `.mes`, the label is the wavelength id.\n- To override them, `Advanced` \u2192 `Channels` is ordered by the **instrument channel number**: element 0 is `Ch1`, element 1 is `Ch2`, over the full range the protocol declares \u2014 not one entry per channel visible in the output. It must be at least as long as the highest channel the acquisition uses; the error names that number if it is too short.\n- One list covers the whole acquisition, projection plates included: each projection algorithm carries its own channel numbers, so the entries are split across plates rather than repeated in each.\n\n### Metadata\n\n- The acquisition's five plate-level vendor files are copied verbatim into `<plate>.zarr/metadata/`: the `.mlf` and `.mrf` under their fixed names, the `.mes`, `.wpi` and `.wpp` under the names recorded in the `.mrf`.\n- Every plate the acquisition produced gets a copy, projection plates included. A file the acquisition does not ship is only a warning, and a name already taken by another acquisition becomes `<stem>_acq{id}<ext>` instead of overwriting.\n\n### Expected inputs\n\nThe following directory structure is expected:\n\n```text\nmy_acquisition/\n\u251c\u2500\u2500 MeasurementData.mlf         # Image measurement records (required)\n\u251c\u2500\u2500 MeasurementDetail.mrf       # Acquisition details and channel info (required)\n\u251c\u2500\u2500 MeasurementProtocol.mes     # Acquisition protocol (optional, sets channel names)\n\u251c\u2500\u2500 NoPlateID.wpi               # Plate definition (optional)\n\u251c\u2500\u2500 10_Greiner_\u03bcClear.wpp       # Plate product (optional)\n\u2514\u2500\u2500 <subdirectories>/\n    \u251c\u2500\u2500 image_001.tif\n    \u2514\u2500\u2500 ...\n```\n\nThe TIFF file paths are referenced inside `MeasurementData.mlf` and can be in subdirectories relative to the acquisition directory. The `.mes`, `.wpi` and `.wpp` filenames vary per acquisition and are read out of the `.mrf` \u2014 the names above are only examples.\n",
       "type": "converter_compound",
-      "executable_non_parallel": "cq3k/convert_cq3k_init_task.py",
+      "executable_non_parallel": "yokogawa/cq3k/convert_cq3k_init_task.py",
       "executable_parallel": "common/image_in_plate_compute_task.py",
       "meta_non_parallel": {
         "cpus_per_task": 1,
