--- src/fractal_uzh_converters/cellvoyager/__init__.py
@@ -1,13 +1,0 @@
-"""Yokogawa CellVoyager to OME-Zarr converters package."""
-
-from fractal_uzh_converters.cellvoyager._utils import CellVoyagerAcquisitionModel
-from fractal_uzh_converters.cellvoyager.api import convert_cellvoyager
-from fractal_uzh_converters.cellvoyager.convert_cellvoyager_init_task import (
-    convert_cellvoyager_init_task,
-)
-
-__all__ = [
-    "CellVoyagerAcquisitionModel",
-    "convert_cellvoyager",
-    "convert_cellvoyager_init_task",
-]
