--- src/fractal_uzh_converters/cq3k/__init__.py
@@ -1,13 +1,0 @@
-"""Yokogawa CQ3K to OME-Zarr converters package."""
-
-from fractal_uzh_converters.cq3k._utils import CQ3KAcquisitionModel
-from fractal_uzh_converters.cq3k.api import convert_cq3k
-from fractal_uzh_converters.cq3k.convert_cq3k_init_task import (
-    convert_cq3k_init_task,
-)
-
-__all__ = [
-    "CQ3KAcquisitionModel",
-    "convert_cq3k",
-    "convert_cq3k_init_task",
-]
