+++ src/fractal_uzh_converters/yokogawa/cq3k/api.py
@@ -1,4 +1,4 @@
-"""Python API for CellVoyager converters."""
+"""Python API for CQ3K converters."""
 
 from ome_zarr_converters_tools import (
     ConverterOptions,
@@ -8,29 +8,29 @@ from ome_zarr_converters_tools import (
 )
 from ome_zarr_converters_tools.fractal import ImageListUpdateDict
 
-from fractal_uzh_converters.cellvoyager._utils import CellVoyagerAcquisitionModel
-from fractal_uzh_converters.cellvoyager.convert_cellvoyager_init_task import (
-    convert_cellvoyager_init_task,
-)
 from fractal_uzh_converters.common.image_in_plate_compute_task import (
     image_in_plate_compute_task,
 )
+from fractal_uzh_converters.yokogawa.cq3k._utils import CQ3KAcquisitionModel
+from fractal_uzh_converters.yokogawa.cq3k.convert_cq3k_init_task import (
+    convert_cq3k_init_task,
+)
 
 
-def convert_cellvoyager(
+def convert_cq3k(
     *,
     zarr_dir: str,
-    acquisitions: list[CellVoyagerAcquisitionModel],
+    acquisitions: list[CQ3KAcquisitionModel],
     converter_options: ConverterOptions | None = None,
     overwrite: OverwriteMode = OverwriteMode.NO_OVERWRITE,
     runner: RunnerType | None = None,
 ) -> list[ImageListUpdateDict]:
-    """Convert a CellVoyager dataset to OME-Zarr.
+    """Convert a CQ3K dataset to OME-Zarr.
 
     Args:
         zarr_dir (str): Directory to store the Zarr files.
-        acquisitions (list[CellVoyagerAcquisitionModel]): List of raw acquisitions to
-            convert to OME-Zarr.
+        acquisitions (list[CQ3KAcquisitionModel]): List of raw acquisitions to convert
+            to OME-Zarr.
         converter_options (ConverterOptions | None): Advanced converter options.
         overwrite (OverwriteMode): Overwrite mode for existing data.
             - "No Overwrite": Do not overwrite existing data.
@@ -52,7 +52,7 @@ def convert_cellvoyager(
         "overwrite": overwrite,
     }
     return exec_compound_task(
-        init_task_fn=convert_cellvoyager_init_task,
+        init_task_fn=convert_cq3k_init_task,
         compute_task_fn=image_in_plate_compute_task,
         init_task_kwargs=init_task_kwargs,
         runner=runner,
